feat: validacao de usuario ocorrendo em handleAdminDeleteUser em adminController.js

+++ src/controllers/adminController.js
@@ -84,6 +84,13 @@ let handleAdminDeleteUser = async (req, res) => {
     if (!matchingPasswords)
         return res.status(401).json({ error: 'Não autorizado!' });
 
+    let isUserRegistered = await User.findOne({
+        where: { email: userEmail }
+    });
+
+    if (!isUserRegistered)
+        return res.status(400).json({ error: 'Usuario não encontrado!' });
+
     try {
         let deletedUser = await User.destroy({
             where: userEmail
